docs: Correções ortográficas;
fix: Correções ortográficas no BoxInfosToRegister;

fix: Removendo todos os console.logs, que acabaram ficando;

fix: Removendo a maior parte das linhas em branco das funcções, mantidas apenas as que fizeram maior sentido;

refactor: Refatorando 2 botões que acabaram ficando como jsx, agora corrigidos!

refactor: Refatorando 2 2 funções, para coloca-las para o helpers e apenas importa-las;
