+++ backend/src/controllers/login.controller.ts
@@ -7,14 +7,10 @@ const loginController = {
   async post(req: Request, res: Response) {
     const { username } = req.body;
     const data = await usersService.readUserByUsername(username);
-
     const user = {...data};
     const id = data.id;
-
     await loginServices.checkPassword(req.body.password, user.password);
-
     const { userNoPass, token } = await jwtServices.createToken({ ...user, id });
-
     res.status(201).json({ ...userNoPass, token });
   },
 };
