+++ backend/src/middlewares/validateToken.ts
@@ -2,10 +2,8 @@ import { NextFunction, Request, Response } from 'express';
 import jwtServices from '../helpers/jwt.services';
 
 const userToken = async (req: Request, _res: Response, next: NextFunction) => {
-
   const auth = await jwtServices.verifyExistsToken(String(req.headers.authorization));
   await jwtServices.decodeToken(auth);
-
   next();
 };
