+++ docker-compose.yml
@@ -36,7 +36,7 @@ services:
     volumes:
       - './backend/:/backend/'
     healthcheck:
-      test: ["CMD", "lsof", "-t", "-i:3007"]  # Caso utilize outra porta interna para o front, altere ela aqui também
+      test: ["CMD", "lsof", "-t", "-i:3007"]  # Caso utilize outra porta interna para o back, altere ela aqui também
       timeout: 10s
       retries: 5
