--- frontend/src/Components/register/BtnNavigateBackHome/BtnNavigateBackHome.jsx
@@ -1,15 +1,0 @@
-import './style/BtnNavigateBackHome.css';
-
-function BtnNavigateBackHome({ name, onClickBack }) {
-  return (
-    <button
-      type="button"
-      className="box-btnBackHome text-btnNameBackHome"
-      onClick={onClickBack}
-    >
-      {name}
-  </button>
-  );
-}
-
-export default BtnNavigateBackHome;
